Fixed #7 (getting stats of possibly non-existing file)

+++ lib/taghelper.py
@@ -135,7 +135,8 @@ class TagVariantLoader(object):
         char_replac_tab = dict(self.__class__.spec_char_replacements)
         translation_table, label_table = load_tag_descriptions(settings.get('session', 'conf_path'), settings.get('session', 'lang'))
 
-        if time.time() - os.stat(path).st_ctime > settings.get_int('cache', 'clear_interval'):
+        if os.path.exists(path) \
+                and time.time() - os.stat(path).st_ctime > settings.get_int('cache', 'clear_interval'):
             os.unlink(path)
 
         if not os.path.exists(path):
